use unique temp folder

+++ Libs/BeatSpiderSharp.Core/SpecialFolders.cs
@@ -1,4 +1,6 @@
-﻿namespace BeatSpiderSharp.Core;
+﻿using Serilog;
+
+namespace BeatSpiderSharp.Core;
 
 public class SpecialFolders: IDisposable
 {
@@ -6,10 +8,12 @@ public class SpecialFolders: IDisposable
     
     public string TempFolder { get; }
 
+    private bool _disposed;
+
     public SpecialFolders()
     {
         DataFolder = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData, Environment.SpecialFolderOption.Create), "BeatSpiderSharp");
-        TempFolder = Path.Combine(DataFolder, "Temp");
+        TempFolder = Path.Combine(DataFolder, "Temp", Path.GetRandomFileName());
 
         Directory.CreateDirectory(DataFolder);
         
@@ -23,7 +27,16 @@ public class SpecialFolders: IDisposable
     
     public void Dispose()
     {
+        if (_disposed) return;
+        _disposed = true;
         GC.SuppressFinalize(this);
-        Directory.Delete(TempFolder, true);
+        try
+        {
+            Directory.Delete(TempFolder, true);
+        }
+        catch (Exception e)
+        {
+            Log.Error(e, "Could not delete temporary folder");
+        }
     }
 }
